requirements.txt
## The following requirements were added by pip freeze:
anyio==3.7.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
dnspython==2.4.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pymongo==4.4.1
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzlocal==5.0.1
